feat(guid): new blueprint
