+++ indexd/app.py
@@ -3,7 +3,8 @@ from .bulk.blueprint import blueprint as indexd_bulk_blueprint
 from .index.blueprint import blueprint as indexd_index_blueprint
 from .alias.blueprint import blueprint as indexd_alias_blueprint
 from .dos.blueprint import blueprint as indexd_dos_blueprint
-from .drs.blueprint import blueprint as indexd_drs_blueprint
+from .drs.blueprint import blueprint as indexd_guid_blueprint
+from .guid.blueprint import blueprint as indexd_drs_blueprint
 from .blueprint import blueprint as cross_blueprint
 
 from indexd.fence_client import FenceClient
@@ -31,6 +32,7 @@ def app_init(app, settings=None):
     app.register_blueprint(indexd_alias_blueprint)
     app.register_blueprint(indexd_dos_blueprint)
     app.register_blueprint(indexd_drs_blueprint)
+    app.register_blueprint(indexd_guid_blueprint)
     app.register_blueprint(cross_blueprint)
     app.register_blueprint(index_urls_blueprint, url_prefix="/_query/urls")
